Send Files via WiFiDIrect (Bi-Directional)
This demo is built with the help of Android native code of WiFiDIrect
but i added some functionality to make this code Bi-Directional. Here
One will be server and other will be client.
Normally client can send file to server, but server can't.
Here server will also be able to send files to client, And transfer of
data will be Bi-Directional.